requirements.txt
transformers>=4.49.0
accelerate
torchdiffeq
librosa>=0.10.2.post1
pyarrow
pandas
bitsandbytes
jieba
cn2an
pypinyin
onnxruntime
x-transformers
pylance
ema-pytorch
prefigure
muq
mutagen
pyopenjtalk
pykakasi
Unidecode
phonemizer
inflect
py3langid
easydict
hydra
omegaconf